Add `rspec-rails` gem
I've opted to use RSpec rather than Minitest as I'm more familiar with
it as a tool so will be able to work more quickly when writing tests.

The version referenced here is '~> 6.0.0', in accordance with RSpec's
documentation - according to their upcoming [versioning strategy][1]
this will match Rails version numbers, but it looks like this hasn't
happened yet. The guidance as this point is to use '~> 6.0.0' even for
Rails 7.x applications.

[1]: https://github.com/rspec/rspec-rails/blob/main/rfcs/versioning-strategy.md