Introduce User model
We're only adding a single, required field initially, `name` which
we're validating on the model and on the DB column.

We don't have a reason to make this more complex i.e. using
`first_name`, `last_name` and we're aiming to imply to users that this
would be their name rather than a 'username', although they'll be free
to enter what they wish.