Add required config for Devise
This is required as devise provides the ability to send emails i.e.
password reset emails, so required action_mailer to have a config
value for `default_url_options`.

This data is not being added for Production here as this is a demo
app, but this would need to be configured later.