Install default Devise views and flash alerts
While I don't know whether we'll want or need all the Devise views,
with our compressed timescale this is the fastest trade-off to get
user auth up and running quickly.

In the future, we can come back and prune what's not needed here.

The Devise docs also indicate that we need to add flashes to our
layout file as it makes use of them. These are using the default
tailwindcss-rails scaffold styling.