Implement Devise for user auth
Devise offers a modular approach to the features it offers. I've added
`confirmable` to the default set provided as we want to prevent
spam/abuse messages so requiring that users validate an email address
can be a useful mitigation for this, albeit one that creates an
additional step for users when registering.