Update Devise to collect 'name' field on user registration
Devise only collects username and password by default, so we need this
additional code too allow the parameter to be allowed ([docs][1]).

We're also removing password confirmation to allow easier/faster
registration. This may increase the risk of typos in passwords getting
through, but hopefully this trade-off is reasonable as there's a
"reset password" function available.

[1]: https://github.com/heartcombo/devise#strong-parameters