Add logo SVG, partial and add them to application layout
We're adding a blank ALT attriute from the logo image as it's not
relevant to the page. Even though we don't have navigation yet, or a
lot of content at the top of the page, we're adding a 'skip to
content' link that's only accessible to screen readers to allow users
to skip whatever gets added and move directly to the content.