Introduce Question
We're adding a 'slug' field here as we want to try and allow users to
write questions before signing up and need a unique, non-sequential
identifier to allow us identify a Question and associate it with a
user once they've registered. For the same reason, the user
association on Question is marked as optional.