Generate slug when a Question is saved
We want to be able to reference a question with a non-sequential ID so
we can assign it to a user after they register.

A trade-off for time here is that we're not rescuing the exception if
the generated slug were to clash with an existing one, an exception
would be raised by the unique index on the column.