Add 'new question' form
This commit introduces our first `turbo_stream` response format, which
Rails handles automatically for us as long as we supply `.turbo_stream`
file indicating how the response should interact with the page. In this
instance, we're replacing the existing form with the updated one
containing the errors. The `dom_id` is required for Turbo to know which
element is being referred to.

The styling here is taken from the scaffolds that the
`tailwindcss-rails` gem provides:
https://github.com/rails/tailwindcss-rails/blob/2b5d9303e466f3333755a4226f4a34deec826339/lib/generators/tailwindcss/scaffold/templates/_form.html.erb.tt