Add questions#show action and view
This commit adds `to_param` to `Question`. I considered solutions such
as the `friendly_id` gem in order to provide additional helpers, but
this simple implementation will be fine for now.

This is an initial implementation of the #show view and only contains
the body with no additional formatting or content. We'll add those in
upcoming commits.