Add redirect logic after creating a question
If a user is already signed in, we can associate that user with the
question being asked. If no user exists, we want to redirect the user
to the registration page, however we want to pass along the `slug` of
the question that's been created so we can associate the user as soon
as they create their account (or, in the future, sign in). We're doing
this by passing the slug as a param - this isn't the most resilient
solution as it won't help if the user opens multiple tabs etc, but
that's a trade-off we're making for this demo app.