Update registrations controller to association questions with users
In a previous commit, we added logic to redirect a user to the
registration form after asking a question if they weren't already
signed in.

This commit adds logic to the registration controller to associate the
user with the question that's been created.

If no `question` param is present i.e. the user hasn't just asked a
question, then for the moment we're redirecting back to the home page
as we don't have anywhere else for now.