Add `question` field to the user registration form
In previous commits, we updated the questions controller to redirect
to the registration form if the user had not yet signed up. This
commit adds a hidden field to the registration form that will pass
this value to the registrations controller.

Note that we're using `hidden_field_tag` rather than `f.hidden_field`
as the field is not part of the `user` and we want it to generate a
param outside of the hash of `user` params.