Add questions#index action and related views
Note that we're limiting the number of questions we're showing on the
index view to the most recent 100 as we don't have time to implement
pagination of any sort.

We're also loading the user association in the same query as we know
we're going to want to display user information for each question.

As we're allowing questions to be created without a user, we're also
only loading questions here that have a user. If we were to implement
the 'only show questions from confirmed users' functionality, we'd
need to filter that criteria here too.