Introduce QuestionResponse model
This model will represent users answers/responses to questions. We've
followed the same mechanism used on the Question model of having a
'slug' in order to allow people to respond to a question before
creating an account.

The slug provides a non-sequential reference to the created record
that we can use to associate the user once they've created an account.