Add QuestionResponsesController with #show and #create actions
Users will be adding question responses from the question view, so
we're not providing a show view, instead redirecting back to the
question view.

As we've done with creating questions, we're once again redirecting
users to the registration form if they create a question_response
without an account. This allows us to associate the user with their
question_response after they've registered.