Add question responses to question show view including form
In previous commits we added a controller for question responses and
updated the user registration logic to handle users that were not
signed in. This means in this commit we can now display the question
responses and a form to add new responses directly under the question.

We're arbitrarily limiting the number of responses displayed to the
most recent 500 to prevent the page from becoming overwhelmed by the
number responses, but in the future we'd need to implement some form
of pagination.

As we allow users to create responses before registering as an
improved user experience, we're also only loading responses that have
an associated user.