Introduce Vote model
We want users to be able to vote on question_responses so they can
bubble up to the top of the order.

We're introducing the vote record as a polymorphic association in case
we want to later allow users to vote on other things, we potentially
won't need to add additional models.