Add votes controller with #create and #destroy actions
We're going to be using Turbo to interact with these actions, so we're
specifying the format of the request as :turbo_stream in the specs.
Rails/Turbo does provide a graceful fallback if Turbo isn't working
i.e. no JS available on the client, but I'm starting to run low on
time, so not implementing this is a trade-off we're going to need to
make for time.