Add vote buttons to question responses
I'm still learning about Turbo, so I'm not sure of the most
performant/efficient way to structure the responses, so I've gone with
the approach of giving each one it's own partial to respond with and
then used the same partial in the original Turbo frame.

We're loading the votes in the controller and then using a `.find` in
the view to see if the current user has a vote. As we limited the
number of question responses that can load, there's a cap on how large
the `@votes` collection can get, but we could update this to use a
method where the implementation isn't in the view i.e.
`User.has_voted_for?(voteable:)` etc.