Sort question responses by the number of votes each has received
At scale, this query may become less performant, so we may want to
implement some sort of caching mechanism, but for the demo we're going
to leave it as a DB query.

Alternatives for this approach could be: a counter-cache column for
the number of votes on the question response, a separate data store
with the number of votes i.e. a separate DB table, redis etc, a
background process to calculate totals and others. The approach would
vary depending on requirements.